UBIFS: fix ia64 compilation warning
For some reasongs %llu gives warning when the argument is le64_to_cpu():

fs/ubifs/replay.c: In function `replay_log_leb':
fs/ubifs/replay.c:807: warning: long long unsigned int format, __le64 arg (arg 6)

Pointed out to by Andrew Morton.

+++ fs/ubifs/recovery.c
@@ -772,7 +772,8 @@ static int get_cs_sqnum(struct ubifs_info *c, int lnum, int offs,
 	}
 	if (le64_to_cpu(cs_node->cmt_no) != c->cmt_no) {
 		dbg_err("CS node cmt_no %llu != current cmt_no %llu",
-			le64_to_cpu(cs_node->cmt_no), c->cmt_no);
+			(unsigned long long)le64_to_cpu(cs_node->cmt_no),
+			c->cmt_no);
 		goto out_err;
 	}
 	*cs_sqnum = le64_to_cpu(cs_node->ch.sqnum);
